@@ -10,12 +10,14 @@
     }
   ],
   "require": {
-    "php": ">=7.0",
-    "ext-mbstring": "*",
+    "php": ">=8.0",
+    "ext-zlib": "*",
     "predis/predis": "^1.1.1"
   },
   "require-dev": {
-    "phpunit/phpunit": "^6.2"
+    "phpunit/phpunit": "^9.5",
+    "friendsofphp/php-cs-fixer": "^3.8",
+    "phpstan/phpstan": "^1.5"
   },
   "autoload": {
     "psr-4": {
@@ -26,5 +28,10 @@
     "psr-4": {
       "B1rdex\\PredisCompressible\\Tests\\": "tests/"
     }
+  },
+  "scripts": {
+    "test": "vendor/bin/phpunit tests/",
+    "fix-cs": "vendor/bin/php-cs-fixer fix --rules=@PSR12:risky,@PHP80Migration:risky,@PHPUnit84Migration:risky --allow-risky=yes .",
+    "sca": "vendor/bin/phpstan --memory-limit=-1 analyse --level=max src/ tests/"
   }
 }
