@@ -6,25 +6,11 @@ namespace B1rdex\PredisCompressible\Compressor;
 
 class ConditionalCompressorWrapper implements CompressorInterface
 {
-    /**
-     * @var int
-     */
-    private $bytesThreshold;
-    /**
-     * @var \B1rdex\PredisCompressible\Compressor\CompressorInterface
-     */
-    private $compressor;
-
-    public function __construct(int $bytesThreshold, CompressorInterface $compressor)
+	public function __construct(private int $bytesThreshold, private CompressorInterface $compressor)
     {
-        $this->bytesThreshold = $bytesThreshold;
-        $this->compressor = $compressor;
-    }
+	}
 
-    /**
-     * {@inheritdoc}
-     */
-    public function compress($data)
+    public function compress(mixed $data): mixed
     {
         if ($this->shouldCompress($data)) {
             return $this->compressor->compress($data);
@@ -33,26 +19,20 @@ class ConditionalCompressorWrapper implements CompressorInterface
         return $data;
     }
 
-    private function shouldCompress($data): bool
+    private function shouldCompress(mixed $data): bool
     {
         if (!\is_string($data)) {
             return false;
         }
 
-        return \mb_strlen($data, 'US-ASCII') > $this->bytesThreshold;
+        return strlen($data) > $this->bytesThreshold;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function isCompressed($data): bool
+    public function isCompressed(mixed $data): bool
     {
         return $this->compressor->isCompressed($data);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function decompress(string $data): string
     {
         return $this->compressor->decompress($data);
