@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace B1rdex\PredisCompressible\Compressor;
 
-interface CompressorException
+interface CompressorException extends \Throwable
 {
 }
