@@ -7,31 +7,21 @@ namespace B1rdex\PredisCompressible\Compressor;
 interface CompressorInterface
 {
     /**
-     * Returns compressed string or original string if its size is less than threshold
+     * Returns a compressed string or original data if its not compressible
      *
-     * @param mixed $data
-     *
-     * @return string|mixed Gzip string or original data
+     * @return string|mixed A compressed string or original data
      */
-    public function compress($data);
+    public function compress(mixed $data): mixed;
 
     /**
-     * Checks if $data is gzipped string
-     *
-     * @param mixed $data
-     *
-     * @return bool
+     * Checks if $data is a compressed string
      */
-    public function isCompressed($data): bool;
+    public function isCompressed(mixed $data): bool;
 
     /**
      * Returns original string from compressed $data
      *
-     * @param string $data
-     *
-     * @return string
-     *
-     * @throws \B1rdex\PredisCompressible\Compressor\CompressorException
+     * @throws CompressorException
      */
     public function decompress(string $data): string;
 }
