@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sp\Tests\PredisCompress;
+namespace B1rdex\PredisCompressible;
 
 use B1rdex\PredisCompressible\Compressor\CompressorException;
 use B1rdex\PredisCompressible\Compressor\ConditionalCompressorWrapper;
@@ -17,13 +17,14 @@ class ConditionalCompressorWrapperTest extends TestCase
     /**
      * @test
      */
-    public function it_should_work()
+    public function it_should_work(): void
     {
         $sut = new ConditionalCompressorWrapper(5, new GzipCompressor());
 
         $data = 'some text string';
 
         $compressed = $sut->compress($data);
+		assert(is_string($compressed));
         $this->assertStringStartsNotWith($data, $compressed);
 
         $decompressed = $sut->decompress($compressed);
